requirements.txt
Flask==2.3.3
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
dnspython>=2.2.0
langdetect==1.0.9
Flask-APScheduler==1.13.0
